requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
whitenoise==4.1.2
